Rebuilt dev environment and added a few updates.
I ran into a problem with a VirtualBox update
and had to rebuild my dev environment. Worked
out for the better anyway.

I udpaded the format of the Config object and
json file, and started making copies of the
HTML from the existing web site so that I can
generate templates for the new implementation.

I also generated a fluxspellsconfig.json file,
but that will not be committed to the version
control repository. A dumby version may be
created eventually to give potential learners
an example.