The 404 Error is gone, but...
I figured out that the reason for the 404 errors was that
I needed to tellt he listener which router to use.

Now, it is kind of working, but the base template that I
borrowed from my other project turns out to have way too
many variable requests related to that project.

However, it is time to sleep, so I will recreate the base
template in the morning. Since I am doing that at this point
anyway, I will base it off of the templates used on the
django version of this project. That way I don't have to
redo as much later.