Milestone 3 Is Complete.
The new base.html template and an initial index.html template
have been hashed out and are working. The Index template is
basically put together but cannot be 100% finished until all
of the other pieces are complete. It will depend on data from
every part of the app. For now it has placeholders for the
main parts that will be displayed.