Started work on User related functionality.
Started work on the User object. Started by copying it over
from another project and starte tweaking it for this one.
I copied in a super basic "crypto" package for handling the
password hashing and login functionality. This is not a high
security project at this point. The added objects still need
a bit of tweaking before I start work on the web forms.