type  CreateUserForm struct
I have added the CreateUserForm struct to facilitate passing the information to the CreateUserFromForm function.
At this time, all that is needed is the email address and the password.  Initially, form validation will be driven by the
handler function, so we do  not need to pass the admin token or password2 to the function that creates the user.